+++ src/astro/sql/operators/agnostic_save_file.py
@@ -22,22 +22,12 @@ import boto3
 import pandas as pd
 from airflow.hooks.base import BaseHook
 from airflow.models import BaseOperator, DagRun, TaskInstance
-
-try:
-    from airflow.providers.google.cloud.hooks.bigquery import BigQueryHook
-except ModuleNotFoundError:
-    from astro.utils.schema_util import RaiseException as BigQueryHook
-
-try:
-    from airflow.providers.postgres.hooks.postgres import PostgresHook
-except ModuleNotFoundError:
-    from astro.utils.schema_util import RaiseException as PostgresHook
-
 from smart_open import open
 
 from astro.sql.operators.temp_hooks import TempSnowflakeHook
 from astro.sql.table import Table
 from astro.utils.cloud_storage_creds import gcs_client, s3fs_creds
+from astro.utils.dependencies import BigQueryHook, PostgresHook
 from astro.utils.schema_util import get_table_name
 from astro.utils.task_id_helper import get_task_id
