+++ src/astro/utils/postgres_transform.py
@@ -15,13 +15,10 @@ limitations under the License.
 """
 import inspect
 
-try:
-    from airflow.providers.postgres.hooks.postgres import PostgresHook
-except ModuleNotFoundError:
-    from astro.utils.schema_util import RaiseException as PostgresHook
 from psycopg2.extensions import AsIs
 
 from astro.sql.table import Table
+from astro.utils.dependencies import PostgresHook
 
 
 def parse_template(sql):
